+++ hspec-core/test/Test/Hspec/Core/ExampleSpec.hs
@@ -26,8 +26,8 @@ spec = do
       it "returns Success on True" $ do
         evaluateExample True `shouldReturn` H.Success
 
-      it "returns Fail on False" $ do
-        evaluateExample False `shouldReturn` H.Fail Nothing ""
+      it "returns Failure on False" $ do
+        evaluateExample False `shouldReturn` H.Failure Nothing ""
 
       it "propagates exceptions" $ do
         evaluateExample (error "foobar" :: Bool) `shouldThrow` errorCall "foobar"
@@ -36,8 +36,8 @@ spec = do
       it "returns Success if all expectations hold" $ do
         evaluateExample (23 `shouldBe` (23 :: Int)) `shouldReturn` H.Success
 
-      it "returns Fail if an expectation does not hold" $ do
-        H.Fail _ msg <- evaluateExample (23 `shouldBe` (42 :: Int))
+      it "returns Failure if an expectation does not hold" $ do
+        H.Failure _ msg <- evaluateExample (23 `shouldBe` (42 :: Int))
         msg `shouldEndWith` "expected: 42\n but got: 23"
 
       it "propagates exceptions" $ do
@@ -66,12 +66,12 @@ spec = do
       it "returns Success if property holds" $ do
         evaluateExample (property $ \n -> n == (n :: Int)) `shouldReturn` H.Success
 
-      it "returns Fail if property does not hold" $ do
-        H.Fail _ _ <- evaluateExample $ property $ \n -> n /= (n :: Int)
+      it "returns Failure if property does not hold" $ do
+        H.Failure _ _ <- evaluateExample $ property $ \n -> n /= (n :: Int)
         return ()
 
       it "shows what falsified it" $ do
-        H.Fail _ r <- evaluateExample $ property $ \ (x :: Int) (y :: Int) -> (x == 0 && y == 1) ==> False
+        H.Failure _ r <- evaluateExample $ property $ \ (x :: Int) (y :: Int) -> (x == 0 && y == 1) ==> False
         r `shouldBe` intercalate "\n"  [
             "Falsifiable (after 1 test): "
           , "0"
@@ -90,7 +90,7 @@ spec = do
         readIORef ref `shouldReturn` 2000
 
       it "pretty-prints exceptions" $ do
-        H.Fail _ r <- evaluateExample $ property (\ (x :: Int) -> (x == 0) ==> (E.throw (E.ErrorCall "foobar") :: Bool))
+        H.Failure _ r <- evaluateExample $ property (\ (x :: Int) -> (x == 0) ==> (E.throw (E.ErrorCall "foobar") :: Bool))
         r `shouldBe` intercalate "\n" [
 #if MIN_VERSION_QuickCheck(2,7,0)
             "uncaught exception: ErrorCall (foobar) (after 1 test)"
@@ -102,7 +102,7 @@ spec = do
 
       context "when used with shouldBe" $ do
         it "shows what falsified it" $ do
-          H.Fail _ r <- evaluateExample $ property $ \ (x :: Int) (y :: Int) -> (x == 0 && y == 1) ==> 23 `shouldBe` (42 :: Int)
+          H.Failure _ r <- evaluateExample $ property $ \ (x :: Int) (y :: Int) -> (x == 0 && y == 1) ==> 23 `shouldBe` (42 :: Int)
           r `shouldStartWith` "Falsifiable (after 1 test): \n"
           r `shouldEndWith` intercalate "\n" [
               "expected: 42"