+++ hspec-core/test/Test/Hspec/Core/FormattersSpec.hs
@@ -20,11 +20,11 @@ testSpec :: H.Spec
 testSpec = do
   H.describe "Example" $ do
     H.it "success"    (H.Success)
-    H.it "fail 1"     (H.Fail Nothing "fail message")
+    H.it "fail 1"     (H.Failure Nothing "fail message")
     H.it "pending"    (H.pendingWith "pending message")
-    H.it "fail 2"     (H.Fail Nothing "")
+    H.it "fail 2"     (H.Failure Nothing "")
     H.it "exceptions" (undefined :: H.Result)
-    H.it "fail 3"     (H.Fail Nothing "")
+    H.it "fail 3"     (H.Failure Nothing "")
 
 spec :: Spec
 spec = do