+++ hspec-core/test/Test/Hspec/Core/RunnerSpec.hs
@@ -424,7 +424,7 @@ spec = do
 
     it "does not let escape error thunks from failure messages" $ do
       r <- silence . H.hspecResult $ do
-        H.it "some example" (H.Fail Nothing $ "foobar" ++ undefined)
+        H.it "some example" (H.Failure Nothing $ "foobar" ++ undefined)
       r `shouldBe` H.Summary 1 1
 
     it "runs specs in parallel" $ do