+++ CHANGES.markdown
@@ -1,3 +1,6 @@
+## Changes next
+  - Rename `Fail` constructor of `Result`-type to `Failure`
+
 ## Changes in 2.3.2
   - Compatibility with HUnit 1.5